flask_consumer/requirements.txt
kafka-python
six
requests